Temporarily update task polling system for Pulp 2
Temporarily update task polling system to address Pulp 2 issue.
Pulp-smash was polling at a 0.3s rate that is occasionally returning
a task "COMPLETE" but the values for the task had yet to be reset.

Adjust the task system to use the previous value of 2s for Pulp 2.

See: https://pulp.plan.io/issues/4795