Add local flag to CLI client
Add local flag to CLI client. This will force the client to run the
CLI commands in the same box that pulp-smash is installed. This will
remove the need to call subprocess to run commands locally.
Besides that, it will allow to reuse the tooling already in
place for the CLI client.