`download_content_unit` now uses smash config
The `pulp_smash.pulp3.utils:download_content_unit` method now forms the
content url to fetch from using the pulp-smash config. Previously this
was formed from the distribution.base_url, but the issue is that
directly users back to nginx/apache which is not how pulp-smash is
intented to test Pulp.

Instead pulp-smash should fetch content directly from the content app
whose port, e.g. 24816, is in the pulp-smash config.