Consolidate some task monitoring functions
There are a lot of use cases for needing parts of the task output that
are not created resources. It's easy enough to get the created
resources if you do want them that we should just return the raw task
output.

monitor_task_groups was defined in pulpcore, but it really should be
defined here along with monitor_task.

[noissue]