Adds support for managing services using s6
The ServiceManager can now stop, start, and restart services inside container that use
s6 to manage services.

The restart command sends a SIGKILL to the service and relies on s6 to restart the service.
This was needed because some services such as PostgreSQL shut down gracefully when SIGTERM
is sent. The graceful shutdown does not happen until all connections are closed by clients.
When testing Pulp, a graceful shutdown is not needed.

[noissue]